refactor: move log into its own class
