+++ parea/cookbook/tracing_with_agent.py
@@ -6,7 +6,8 @@ from dotenv import load_dotenv
 
 from parea import Parea
 from parea.helpers import to_date_and_time_string
-from parea.schemas.models import Completion, CompletionResponse, FeedbackRequest, LLMInputs, Message, ModelParams, Role
+from parea.schemas.models import Completion, CompletionResponse, FeedbackRequest
+from parea.schemas.log import Role, Message, ModelParams, LLMInputs
 from parea.utils.trace_utils import get_current_trace_id, trace
 
 load_dotenv()
