+++ parea/cookbook/tracing_without_deployed_prompt.py
@@ -4,7 +4,8 @@ from datetime import datetime
 from dotenv import load_dotenv
 
 from parea import Parea
-from parea.schemas.models import Completion, CompletionResponse, FeedbackRequest, LLMInputs, Message, ModelParams
+from parea.schemas.models import Completion, CompletionResponse, FeedbackRequest
+from parea.schemas.log import Message, ModelParams, LLMInputs
 from parea.utils.trace_utils import get_current_trace_id, trace
 
 load_dotenv()
