+++ parea/schemas/models.py
@@ -1,47 +1,9 @@
-from typing import Any, Dict, List, Optional, Union
+from typing import Any, List, Optional
 
-from enum import Enum
 
 from attrs import define, field, validators
 
-
-class Role(str, Enum):
-    user = "user"
-    assistant = "assistant"
-    system = "system"
-    example_user = "example_user"
-    example_assistant = "example_assistant"
-
-
-@define
-class Message:
-    content: str
-    role: Role = Role.user
-
-    def to_dict(self) -> dict[str, str]:
-        return {
-            "content": self.content,
-            "role": str(self.role),
-        }
-
-
-@define
-class ModelParams:
-    temp: float = 1.0
-    top_p: float = 1.0
-    frequency_penalty: float = 0.0
-    presence_penalty: float = 0.0
-    max_length: Optional[int] = None
-
-
-@define
-class LLMInputs:
-    model: Optional[str] = None
-    provider: Optional[str] = None
-    model_params: Optional[ModelParams] = None
-    messages: Optional[List[Message]] = None
-    functions: Optional[List[Any]] = None
-    function_call: Optional[Union[str, dict[str, str]]] = None
+from parea.schemas.log import LLMInputs, Log
 
 
 @define
@@ -119,19 +81,6 @@ class FeedbackRequest:
 class NamedEvaluationScore:
     name: str
     score: float = field(validator=[validators.ge(0), validators.le(1)])
-
-
-@define
-class Log:
-    configuration: LLMInputs = LLMInputs()
-    inputs: Optional[Dict[str, str]] = None
-    output: Optional[str] = None
-    target: Optional[str] = None
-    latency: Optional[float] = 0.0
-    input_tokens: Optional[int] = 0
-    output_tokens: Optional[int] = 0
-    total_tokens: Optional[int] = 0
-    cost: Optional[float] = 0.0
 
 
 @define
